pyproject.toml
[project]
name = "langchain-azure-ai"
version = "0.1.0"
description = "A PoC application demonstrating document-based chat using Azure AI Foundry AI Agent Service and Chainlit"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "azure-ai-agents>=1.0.2",
    "azure-identity>=1.23.0",
    "chainlit>=2.6.0",
]
